+++ src/mobi/hsz/idea/gitignore/projectView/HideIgnoredFilesTreeStructureProvider.java
@@ -1,4 +1,28 @@
-package mobi.hsz.idea.gitignore.hideignoredfiles;
+/*
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2017 hsz Jakub Chrzanowski <jakub@hsz.mobi>
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy
+ * of this software and associated documentation files (the "Software"), to deal
+ * in the Software without restriction, including without limitation the rights
+ * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+ * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+ * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+ * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
+ * SOFTWARE.
+ */
+
+package mobi.hsz.idea.gitignore.projectView;
 
 import com.intellij.ide.projectView.TreeStructureProvider;
 import com.intellij.ide.projectView.ViewSettings;
@@ -17,14 +41,15 @@ import java.util.ArrayList;
 import java.util.Collection;
 
 /**
- * Extension for the Project Tree View that provides it with the ability to hide from
- * it files or folders that had been marked as ignored
+ * Extension for the {@link TreeStructureProvider} that provides the ability to hide ignored files
+ * or directories in the project tree view.
  *
  * @author Maximiliano Najle <maximilianonajle@gmail.com>
  * @since 1.7
  */
 public class HideIgnoredFilesTreeStructureProvider implements TreeStructureProvider {
 
+    private final IgnoreSettings ignoreSettings = IgnoreSettings.getInstance();
     private final Project project = ProjectManager.getInstance().getOpenProjects()[0];
     private final IgnoreManager ignoreManager = IgnoreManager.getInstance(project);
 
@@ -34,16 +59,17 @@ public class HideIgnoredFilesTreeStructureProvider implements TreeStructureProvi
                                                @NotNull Collection<AbstractTreeNode> children,
                                                ViewSettings settings) {
 
-        if (!IgnoreSettings.getInstance().shouldHideIgnoredFilesOnProjectView()) {
+        if (!ignoreSettings.isHideIgnoredFiles()) {
             return children;
         }
 
-        ArrayList<AbstractTreeNode> nodes = new ArrayList<AbstractTreeNode>();
+        final ArrayList<AbstractTreeNode> nodes = new ArrayList<AbstractTreeNode>();
         for (AbstractTreeNode child : children) {
             if (!isAnIgnoredFile(child)) {
                 nodes.add(child);
             }
         }
+        
         return nodes;
     }
 
@@ -55,16 +81,15 @@ public class HideIgnoredFilesTreeStructureProvider implements TreeStructureProvi
 
     private boolean isAnIgnoredFile(AbstractTreeNode node) {
         try {
-            VirtualFile file = ((PsiFileNode) node).getVirtualFile();
+            final VirtualFile file = ((PsiFileNode) node).getVirtualFile();
             return ignoreManager.isFileIgnored(file);
         } catch (Exception e) {
             try {
-                VirtualFile file = ((PsiDirectoryNode) node).getVirtualFile();
+                final VirtualFile file = ((PsiDirectoryNode) node).getVirtualFile();
                 return ignoreManager.isFileIgnored(file);
             } catch (Exception ex) {
                 return false;
             }
         }
     }
-
 }