+++ src/mobi/hsz/idea/gitignore/settings/IgnoreSettings.java
@@ -64,7 +64,7 @@ public class IgnoreSettings implements PersistentStateComponent<Element>, Listen
         OUTER_IGNORE_RULES("outerIgnoreRules"), OUTER_IGNORE_WRAPPER_HEIGHT("outerIgnoreWrapperHeight"),
         INSERT_AT_CURSOR("insertAtCursor"), ADD_UNVERSIONED_FILES("addUnversionedFiles"), VERSION("version"),
         STARRED_TEMPLATES("starredTemplates"), UNIGNORE_ACTIONS("unignoreActions"),
-        HIDE_IGNORED_FILES_ON_PROJECT_VIEW("hideIgnoredFilesOnProjectView");
+        HIDE_IGNORED_FILES("hideIgnoredFiles");
 
         private final String key;
 
@@ -109,7 +109,7 @@ public class IgnoreSettings implements PersistentStateComponent<Element>, Listen
     /** Enable unignore actions in context menus. */
     private boolean unignoreActions = true;
 
-    /* Hide ignored files or folder in the project tree view*/
+    /** Hide ignored files or folder in the project tree view. */
     private boolean hideIgnoredFiles = false;
 
     /** Starred templates. */
@@ -159,7 +159,7 @@ public class IgnoreSettings implements PersistentStateComponent<Element>, Listen
         element.setAttribute(KEY.VERSION.toString(), version);
         element.setAttribute(KEY.STARRED_TEMPLATES.toString(), StringUtil.join(starredTemplates, Constants.DOLLAR));
         element.setAttribute(KEY.UNIGNORE_ACTIONS.toString(), Boolean.toString(unignoreActions));
-        element.setAttribute(KEY.HIDE_IGNORED_FILES_ON_PROJECT_VIEW.toString(), Boolean.toString(hideIgnoredFiles));
+        element.setAttribute(KEY.HIDE_IGNORED_FILES.toString(), Boolean.toString(hideIgnoredFiles));
 
         Element languagesElement = new Element(KEY.LANGUAGES.toString());
         for (Map.Entry<IgnoreLanguage, TreeMap<IgnoreLanguagesSettings.KEY, Object>> entry : languagesSettings.entrySet()) {
@@ -222,8 +222,8 @@ public class IgnoreSettings implements PersistentStateComponent<Element>, Listen
         value = element.getAttributeValue(KEY.STARRED_TEMPLATES.toString());
         if (value != null) setStarredTemplates(StringUtil.split(value, Constants.DOLLAR));
 
-        value = element.getAttributeValue(KEY.HIDE_IGNORED_FILES_ON_PROJECT_VIEW.toString());
-        hideIgnoredFiles = (value != null) && Boolean.parseBoolean(value);
+        value = element.getAttributeValue(KEY.HIDE_IGNORED_FILES.toString());
+        if (value != null) hideIgnoredFiles = Boolean.parseBoolean(value);
 
         Element languagesElement = element.getChild(KEY.LANGUAGES.toString());
         if (languagesElement != null) {
@@ -370,24 +370,22 @@ public class IgnoreSettings implements PersistentStateComponent<Element>, Listen
     }
 
     /**
-     * Check if ignored files should be hidden in the project tree view
+     * Check if ignored files should be hidden in the project tree view.
      *
      * @return true if the files should be ignored and false if they should be showed
      */
-    public boolean shouldHideIgnoredFilesOnProjectView() {
+    public boolean isHideIgnoredFiles() {
         return hideIgnoredFiles;
     }
 
     /**
      * Changes the configuration to determine if ignored files should be hidden in the project tree view or not
      *
+     * @param hideIgnoredFiles should hide ignored files
      */
-    public void toggleIgnoredFilesVisibility() {
-        boolean newValue = !this.hideIgnoredFiles;
-
-        this.notifyOnChange(KEY.HIDE_IGNORED_FILES_ON_PROJECT_VIEW,
-                this.hideIgnoredFiles, newValue);
-        this.hideIgnoredFiles = newValue;
+    public void setHideIgnoredFiles(boolean hideIgnoredFiles) {
+        this.notifyOnChange(KEY.HIDE_IGNORED_FILES, this.hideIgnoredFiles, hideIgnoredFiles);
+        this.hideIgnoredFiles = hideIgnoredFiles;
 
         ProjectView.getInstance(ProjectManager.getInstance().getOpenProjects()[0]).refresh();
     }