+++ CHANGELOG.md
@@ -6,11 +6,12 @@
 
 **Implemented enhancements:**
 
-- Hide ignored files in the project tree view
+- Hide ignored files and directories in the project tree view
 
 **Fixed bugs:**
 
 - 
+
 ## [v1.6](https://github.com/hsz/idea-gitignore/tree/v1.6) (2016-11-23)
 
 [Full Changelog](https://github.com/hsz/idea-gitignore/compare/v1.5...v1.6)